CPU MegaList downloaded
I just copied the response of the server since I was not able to reverse-engineer website's rules of data fetching. I might need it in the future, but I'm not sure if the data is updated there. The structure of the file is also up to debate, since both the id and the name should be unique, yet they are not the keys of the whole list, but rather keys of inner "cpu" dictionaries, which doesn't help in searching at all