tests! and more
I decided to introduce tests early in the development so that I don't have to deal with that much debugging in the future. The code now works for all codes available in the checkup table (I don't think I'll use any apart from those). Test automation now checks whether the list of all manufacturers from the "database" (aka .json file) matches the result of web scraper fetching. Finally removed the silliest of classes -- "Filters". it was a weird idea since the beginning, but it's gone now. The same applies to human_notes()