MBDB silliness reverts
Jesus my code was so ass back in the day... I was parsing the amount of results (like, 20 per page) instead of parsing the amount of pages... I was wondering why my request times were so long for MBDB.com, so I did all of those perf_counters (which I'm going to remove in the next commit) to see the issue. Now parsing takes 1 seconds instead of 11. i think it's a big W