provantage filters, ErrorMessage & compatibility (school April 12th, 2024)
Started working on fetching and using filters for provantage components (RAM and PSU). For now, they are chiling in all_jsons/test.json, since some improvements have to be done. Adapted provantage.py to fetch and use further filters (those under "more features" on the website). Changed database to use the new keywords for filters
Changed the order of checks done in compatibility_check.py in the order of "importance" (IMHO) and provided error messages for critical incompatibilities. This could be changed from print() to ErrorMessage() in the future, just for the coolness points
Finally removed the .png file lol