April 16th 2024 (#8)
* Refactored CPU fetching into a table in the html

* small tweaks and tech_p_up search by words

* proper "requirements.txt"

* introducing cpu_search_form part 1

Unfortunately, it's not finished yet. If you try to run it, it gives an error that it requires CSRF
Created a temporary function in class_database to search for CPU by the name
I would have to reformat the db to find a component by component name, not what I have now

* kinda made search for cpu

it's in a very rough condition, but I was able to implement a search of CPU by name using a custom form on a main page
A lot more to come

* last minute changes

* adapting styles and inspections

I want consistency in the coding styles and inspection errors, so I will copy the .xml profiles into a separate folder to save it for the future computers. Hopefully, I won't forget to remove them when deploying

* giving up on inspection profiles

I have no clue how .gitignore excludes patterns, and I don't want to lose time on that, so I'll just carry it on my USB I guess

* decent CPU+GPU search

That was a lot of work, refactoring and debugging, but we finally did it
Changed library from flask-bootstrap to bootstrap-flask (which share THE SAME BLOODY NAME, I'm so mad) to use Bootstrap5. Had to refactor a lot of HTML for that
The search for CPU and GPU by name finally works and parses *all the values* (that was a very silly error that took me a lot to debug)
introduced the .html pages for lists of those components, each with its own set of properties to show
Changed the main page to display the search form
And more, but my parents are going to kill me if I don't help them so that's it for this message

---------

Co-authored-by: school computer <school_computer@gmail.com>